Converted components to typescript.
